finalize maze
@anthonykulis
this is the submission for extra credit. You said it was okay to specify the size of the maze in the maze file, so this program only works on my maze files.
I kept the key listener from my previous version, it seemed a shame to waste that work.
I have switched the input to command line, enter a maze file if you want to, and if you do you can specify the number of monsters as the next arg.